### 2023-04-07 #5348 @editorialbot
[REVIEW]: ExoRad 2.0: The generic point source radiometric model
**Submitting author:** @LorenzoMugnai (<a href="http://orcid.org/0000-0002-9007-9802">Lorenzo V. Mugnai</a>)
**Repository:** https://github.com/ExObsSim/ExoRad2-public
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1.121
