### 2023-04-27 #5348 @skendrew
is there a policy of all contributors being listed as co-author on the paper, or do we leave that up to the lead author/developer? a minor contribution was made by someone who I believe is not a co-author, not sure if that is something to raise?
