### 2023-04-27 ExObsSim/ExoRad2-public#2 @skendrew
help command error
@openjournals/joss-reviews#5348

the README suggests running `exorad -help` for usage instructions but I believe it should be `exorad -h` or `exorad --help`. `exorad -help` returns an error.
