### 2023-04-27 #5348 @dfm
@skendrew — good question! The JOSS authorship policy is [here](https://joss.readthedocs.io/en/latest/submitting.html#authorship), and there isn't a requirement that all contributors are co-authors, or that all authors must have contributed to the codebase. We typically leave the specific authorship choices up to the authors and contributors unless there seems to be any significant omissions! It's common for minor contributions to be considered insufficient for co-authorship.
