### 2023-04-28 #5348 @LorenzoMugnai
Hi. Thank you for noticing :)
In this case, I believe that this is a very minor contribution.
As you can see [here](https://github.com/ExObsSim/ExoRad2-public/pull/1/files), the contributor noticed that I left a directory with my name on it in the configuration file and suggested removing it. I would not consider it a contribution to the code.
