### 2023-05-10 ExObsSim/ExoRad2-public#2 @LorenzoMugnai
Thank you for noticing it! It should be fixed now.
