### 2023-05-23 #5348 @eas342
Sorry I don't know the appropriate place to put comments on the text of the PDF so I'm putting them here @LorenzoMugnai.

- I suggest a different reference to the error in MULTIACCUM. I believe (https://ui.adsabs.harvard.edu/abs/2007PASP..119..768R/abstract) and (https://ui.adsabs.harvard.edu/abs/2016SPIE.9910E..16P/abstract) are better ones to reference than PandExo.
- Figure 2: The noise on the noise budget. I suggest putting in the units as sqrt(texp/hr) since the relative noise should be unitless.
