### 2023-05-23 #5348 @eas342
And the authors may consider citing author telescope simulation software

>  State of the field: Do the authors describe how this software compares to other commonly-used packages?

I'm honestly not too familiar with current telescope software that runs quickly like ExoRad but some examples of simulators I have seen are are Pandeia, WebbPSF, https://github.com/jarronl/pynrc , Code V, Zemax, Phosim (https://www.lsst.org/scientists/simulations/phosim).
