+++ joss/paper.bib
@@ -137,3 +137,42 @@ archivePrefix = {arXiv},
 
 
 
+@ARTICLE{rauscher,
+       author = {{Rauscher}, Bernard J. and {Fox}, Ori and {Ferruit}, Pierre and {Hill}, Robert J. and {Waczynski}, Augustyn and {Wen}, Yiting and {Xia-Serafino}, Wei and {Mott}, Brent and {Alexander}, David and {Brambora}, Clifford K. and {Derro}, Rebecca and {Engler}, Chuck and {Garrison}, Matthew B. and {Johnson}, Thomas and {Manthripragada}, Sridhar S. and {Marsh}, James M. and {Marshall}, Cheryl and {Martineau}, Robert J. and {Shakoorzadeh}, Kamdin B. and {Wilson}, Donna and {Roher}, Wayne D. and {Smith}, Miles and {Cabelli}, Craig and {Garnett}, James and {Loose}, Markus and {Wong-Anglin}, Selmer and {Zandian}, Majid and {Cheng}, Edward and {Ellis}, Timothy and {Howe}, Bryan and {Jurado}, Miriam and {Lee}, Ginn and {Nieznanski}, John and {Wallis}, Peter and {York}, James and {Regan}, Michael W. and {Hall}, Donald N.~B. and {Hodapp}, Klaus W. and {B{\"o}ker}, Torsten and {De Marchi}, Guido and {Jakobsen}, Peter and {Strada}, Paolo},
+        title = "{Detectors for the James Webb Space Telescope Near-Infrared Spectrograph. I. Readout Mode, Noise Model, and Calibration Considerations}",
+      journal = {\pasp},
+     keywords = {Astronomical Instrumentation, Astrophysics},
+         year = 2007,
+        month = jul,
+       volume = {119},
+       number = {857},
+        pages = {768-786},
+          doi = {10.1086/520887},
+archivePrefix = {arXiv},
+       eprint = {0706.2344},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2007PASP..119..768R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@INPROCEEDINGS{pandeia,
+       author = {{Pontoppidan}, Klaus M. and {Pickering}, Timothy E. and {Laidler}, Victoria G. and {Gilbert}, Karoline and {Sontag}, Christopher D. and {Slocum}, Christine and {Sienkiewicz}, Mark J. and {Hanley}, Christopher and {Earl}, Nicholas M. and {Pueyo}, Laurent and {Ravindranath}, Swara and {Karakla}, Diane M. and {Robberto}, Massimo and {Noriega-Crespo}, Alberto and {Barker}, Elizabeth A.},
+        title = "{Pandeia: a multi-mission exposure time calculator for JWST and WFIRST}",
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
+    booktitle = {Observatory Operations: Strategies, Processes, and Systems VI},
+         year = 2016,
+       editor = {{Peck}, Alison B. and {Seaman}, Robert L. and {Benn}, Chris R.},
+       series = {Society of Photo-Optical Instrumentation Engineers (SPIE) Conference Series},
+       volume = {9910},
+        month = jul,
+          eid = {991016},
+        pages = {991016},
+          doi = {10.1117/12.2231768},
+archivePrefix = {arXiv},
+       eprint = {1707.02202},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016SPIE.9910E..16P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+

[…]
