+++ joss/paper.md
@@ -55,7 +55,7 @@ ExoRad 2.0 is a simulator able to accurately predict the telescope performance i
 
 ![ExoRad computes the instrument's optical efficiency (red and orange for the photometer and the spectrometer respectively) by combining the optical elements and the foregrounds between the target star and the telescope. By combining the optical efficiency with the detector quantum efficiency (green and brown for the photometer and the spectrometer respectively), it measures the photon conversion efficiency (black) for each channel. In this example, we report the results for an instrument consisting of a photometer (blue band) and a spectrometer (red band), which is included as a quick-start simulation in the source code package. \label{fig:efficiency}](efficiency.png){height=90%}
 
-ExoRad 2.0 uses simulated PFSs to output the estimated signals on the detector pixels. Different formats of PSFs are supported (such as PAOS [@paos] products) and more can be easily added to the software in the future. If no PSF is provided, ExoRad 2.0 uses a simple Airy PSF. From the signal, ExoRad 2.0 computes the relative noise versus the spectral bins (see \autoref{fig:myTest}). The software also returns the maximum signal on a pixel for each spectral bin and estimates the detector saturation time. The noise sources included in the simulation are not limited to the photon noise arising from the signal. The software includes detector noise as dark current and read noise, and a noise gain factor related to the readout noise and the Multiaccum equation [@rauscher,@pandeia], which is also described in @arielrad. Custom wavelength-dependent noise sources can be included at the instrument or channel level from the input file. The noise output is in units of relative noise on one hour integration time, such that can be easily rescaled to the desired observing time.
+ExoRad 2.0 uses simulated PFSs to output the estimated signals on the detector pixels. Different formats of PSFs are supported (such as PAOS [@paos] products) and more can be easily added to the software in the future. If no PSF is provided, ExoRad 2.0 uses a simple Airy PSF. From the signal, ExoRad 2.0 computes the relative noise versus the spectral bins (see \autoref{fig:myTest}). The software also returns the maximum signal on a pixel for each spectral bin and estimates the detector saturation time. The noise sources included in the simulation are not limited to the photon noise arising from the signal. The software includes detector noise as dark current and read noise, and a noise gain factor related to the readout noise and the Multiaccum equation (@rauscher, @pandeia), which is also described in @arielrad. Custom wavelength-dependent noise sources can be included at the instrument or channel level from the input file. The noise output is in units of relative noise on one hour integration time, such that can be easily rescaled to the desired observing time.
 ExoRad 2.0 can estimate the performance of entire target lists. By analyzing 1000 candidate targets in a 20 minutes time scale, the software allows the validation of different observational strategies.
 
 ![ExoRad produces diagnostic plots to summarise the contributions to the signal and to the noise. In this example, we run the quick-start simulation included in the package, where the instrument consists of a photometer (blue band) and a spectrometer (red band). The figure shows the contribution to the signal on the top panel, where each data point corresponds to a spectral bin computed according to the spectral resolving power indicated in the instrument description. In this example, the main contribution to the signal is the flux from the target stars. Other contributions are considered: `zodi_signal`, referring to the zodiacal emission, and `skyFilter_signal`, which is a custom contribution included in the simulation. `instrument_signal`, which refers to the instrument self-emission, is considered in this example, but its contribution is too small for the figure axis range. The bottom panel shows the noise relative to the signal integrated on a time scale ($t_{int}$) of 1 hour. Other noise contributions arising from the signals are included on top of the photon noise. The noise sources include the dark current, the read noise and three custom noise sources called `gain noise`. \label{fig:myTest}](myTest.png){height=90%}
