+++ joss/paper.bib
@@ -97,21 +97,23 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{twinkle,
-       author = {{Edwards}, Billy and {Rice}, Malena and {Zingales}, Tiziano and {Tessenyi}, Marcell and {Waldmann}, Ingo and {Tinetti}, Giovanna and {Pascale}, Enzo and {Savini}, Giorgio and {Sarkar}, Subhajit},
-        title = "{Exoplanet spectroscopy and photometry with the Twinkle space telescope}",
-      journal = {Experimental Astronomy},
-     keywords = {Twinkle space mission, Exoplanet photometry and spectroscopy, Exoplanet atmospheres, Space telescope, Astrophysics - Earth and Planetary Astrophysics},
-         year = 2019,
-        month = apr,
-       volume = {47},
-       number = {1-2},
-        pages = {29-63},
-          doi = {10.1007/s10686-018-9611-4},
+@INPROCEEDINGS{twinkle,
+       author = {{Stotesbury}, Ian and {Edwards}, Billy and {Lavigne}, Jean-Francois and {Pesquita}, Vasco and {Veilleux}, James and {Windred}, Philip and {Al-Refaie}, Ahmed and {Bradley}, Lawrence and {Ma}, Sushang and {Savini}, Giorgio and {Tinetti}, Giovanna and {Birnstiel}, Til and {Dodson-Robinson}, Sally and {Ercolano}, Barbara and {Feliz}, Dax and {Hernitschek}, Nina and {Holdsworth}, Daniel and {Jiang}, Ing-Guey and {Griffin}, Matt and {Lowson}, Nataliea and {Molaverdikhani}, Karan and {Neilson}, Hilding and {Phillips}, Caprice and {Preibisch}, Thomas and {Sarkar}, Subhajit and {Stassun}, Keivan G. and {Ward-Thompson}, Derek and {Wright}, Duncan and {Yang}, Ming and {Yeh}, Li-Chin and {Zhou}, Ji-Lin and {Archer}, Richard and {Barrathwaj Raman Mohan}, Yoga and {Joshua}, Max and {Tessenyi}, Marcell and {Tennyson}, Jonathan and {Wilcock}, Benjamin},
+        title = "{Twinkle: a small satellite spectroscopy mission for the next phase of exoplanet science}",
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
+    booktitle = {Space Telescopes and Instrumentation 2022: Optical, Infrared, and Millimeter Wave},
+         year = 2022,
+       editor = {{Coyle}, Laura E. and {Matsuura}, Shuji and {Perrin}, Marshall D.},
+       series = {Society of Photo-Optical Instrumentation Engineers (SPIE) Conference Series},
+       volume = {12180},
+        month = aug,
+          eid = {1218033},
+        pages = {1218033},
+          doi = {10.1117/12.2641373},
 archivePrefix = {arXiv},
-       eprint = {1811.08348},
- primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ExA....47...29E},
+       eprint = {2209.03337},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022SPIE12180E..33S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
