+++ joss/paper.bib
@@ -176,5 +176,3 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2016SPIE.9910E..16P},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-
