+++ setup.py
@@ -2,5 +2,4 @@
 from setuptools import setup
 
 if __name__ == "__main__":
-
     setup()
