### 2023-05-26 #5348 @eas342
I made a random list of 1000 stars, and confirmed the performance claim - it took 23 minutes to run on an M1mac.
