### 2023-05-26 #5348 @eas342
@LorenzoMugnai , the documentation looks good! I think the only thing that JOSS requires is a statement of need - I think that could be copied from the JOSS writeup.
