### 2023-05-27 9f4873d
## [2.1.118] - 2023-05-27 ### Fixed - deprecated np.float removed - deprecated quantity interpreted as boolean
