+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.1.118] - 2023-05-27
+### Fixed
+- deprecated np.float removed
+- deprecated quantity interpreted as boolean
+
 ## [2.1.116] - 2023-02-28
 ### Fixed
 - exosim reference removed
@@ -242,6 +247,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Initial release
 
 [Unreleased]: https://github.com/ExObsSim/ExoRad2.0
+[2.1.118]: https://github.com/ExObsSim/ExoRad2-public/compare/v2.1.116...v2.1.118
 [2.1.116]: https://github.com/ExObsSim/ExoRad2-public/compare/v2.1.115...v2.1.116
 [2.1.115]: https://github.com/ExObsSim/ExoRad2-public/compare/v2.1.114...v2.1.115
 [2.1.114]: https://github.com/ExObsSim/ExoRad2-public/compare/v2.1.113...v2.1.114
