+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = exorad
-version = 2.1.116
+version = 2.1.118
 author = Lorenzo V. Mugnai, Enzo Pascale
 author_email = lorenzo.mugnai@uniroma1.it
 description = The generic point source radiometric model
