### 2023-05-27 #5348 @LorenzoMugnai
@eas342 Thank you for your great feedback. I
have updated the documentation, please let me know if it answers the requirements now:

> A statement of need
The authors should clearly state what problems the software is designed to solve, who the target audience is, and its relation to other work.

I am working on the paper too, but I am traveling and it is requiring me some more time.
Thank again!
