+++ joss/paper.bib
@@ -176,3 +176,19 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2016SPIE.9910E..16P},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@MISC{synphot,
+       author = {{STScI Development Team}},
+        title = "{synphot: Synthetic photometry using Astropy}",
+     keywords = {Software},
+ howpublished = {Astrophysics Source Code Library, record ascl:1811.001},
+         year = 2018,
+        month = nov,
+          eid = {ascl:1811.001},
+        pages = {ascl:1811.001},
+archivePrefix = {ascl},
+       eprint = {1811.001},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018ascl.soft11001S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
