+++ joss/paper.md
@@ -65,7 +65,7 @@ Since the early phases of designing and developing instruments, we need fast and
 
 ExoRad 2.0 allows the same level of flexibility and accuracy as ArielRad, but it is now compatible with any photometric or spectroscopic instrument. The software is written following an object-oriented programming paradigm, allowing the user to easily extend the software to include new functionalities. The package includes a default pipeline for the simulation, but it can also be used as a library to build custom simulations. An example of the latter is included in the source code as a Python notebook. The software is compatible with any instrument having single or multiple channels. This allows the user to easily simulate different optical paths or different configurations for the same instrument. The software is also compatible with any target, allowing the user to easily include new target types in the software. The software is continuously updated and improved to include new functionalities and to be compatible with new instruments.
 
-Similar simulators are Pandeia [@pandeia] and synphot [@synphot]. The first is widely used in the community. Pandeia is a Python package developed by the Space Telescope Science Institute (STScI) to simulate the performance of the James Webb Space Telescope (JWST). The software is used by the community as an exposure time calculator to predict the noise on single target observations. synphot is a Python package developed by the STScI to simulate photometric data and spectra, later adapted to the Hubble Space Telescope (HST). Compared to them, ExoRad is easier to use and to adapt to different telescopes, as requires less data as inputs. This feature is particularly useful to design a new observatory. ExoRad is more flexible, as allows the user to easily include new instruments. ExoRad is also more versatile, as allows the user to simulate the performance of an entire target list in a short time scale.
+Similar simulators are Pandeia [@pandeia] and synphot [@synphot]. Pandeia is a Python package developed by the Space Telescope Science Institute (STScI) to mainly simulate the performance of the James Webb Space Telescope (JWST). The software is used by the community as an exposure time calculator to predict the noise on single target observations. synphot is a Python package developed by the STScI to simulate photometric data and spectra, later adapted to the Hubble Space Telescope (HST). Compared to them, ExoRad is easier to use and to adapt to different telescopes, as requires less data as inputs, a feature specifically designed to make ExoRad a valid ally in designing new observatories. ExoRad is more flexible, as allows the user to easily include new instruments. ExoRad is also more versatile, as allows the user to simulate the performance of an entire target list in a short time scale.
  
 
 The ExoRad software is used by the community to design their instruments and validate their performance and select the best targets to optimize the scientific return.
