### 2023-05-30 #5348 @LorenzoMugnai
> And the authors may consider citing other telescope simulation software
>
> > State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> I'm honestly not too familiar with current telescope software that runs quickly like ExoRad but some examples of simulators I have seen are are Pandeia, WebbPSF, https://github.com/jarronl/pynrc , Code V, Zemax, Phosim (https://www.lsst.org/scientists/simulations/phosim).
>
> Or maybe [synphot](https://synphot.readthedocs.io) is more similar to ExoRad

Hi, @eas342 , I updated the manuscript by adding a paragraph (line 82). Please let me know if this is better.
I only referred to Pandeia and synphot, because the others are more optical/PSF simulators, so are not really comparable to ExoRad. But I am open to any comments or suggestions.

I also updated the figures and the citation to the MULTIACCUM equation, as you suggested.
Thank you
