### 2023-06-14 #5348 @skendrew
I have mainly minor comments and apologies for being slow. Thank you for creating this software, having spent a lot of my career working in performance estimation of instruments at various stages of development I agree that having robust and versatile software for this work is very valuable. I like how an optical path can be entirely configured in the input.

I encountered no issues with the installation or functionality of the code. Documentation is well written and clear.

**A few comments & questions for the text of the paper and/or documentation**
* I think it would be very useful if you could include in the text of the paper a short discussion of the use of the code for ground-based vs. space-based instruments. Or perhaps include an example in the code for a ground-based telescope/instrument? I gather it's been written primarily for space-based (or airborne) systems, are there limitations for its use for ground-based system?
* I can see in the payload_example input file that there are a number of options for providing a PSF. The description in the paper and documentation of this is fairly sparse. Can you provide a bit more information on what options are available for providing custom PSFs? are time-varying PSFs supported?
* has the code been validated against any actual working telescopes/instruments? you mention validation against ArielRad and ExoSim; can you elaborate in the paper on how these have been validated against working systems? this would be useful for the reader.
* it would be good to clarify what the min/max wavelengths are that the code can handle. I can see that the use can specify cutoff wavelengths, but presumably there is is some limit?

Thanks!
