### 2023-06-17 #5348 @eas342
@LorenzoMugnai The new manuscript looks good - thank you for making the changes. I would just echo Sarah that it would be nice to verify that the results make sense on a calibrated telescope. Can you provide a simple example, say for HST, JWST, TESS etc. that can be checked against an Exposure Time Calculator or observed data? Other than that, I think it's good to go on my end.
