+++ joss/paper.md
@@ -48,7 +48,8 @@ ExoRad 2.0 has an extensive documentation, available on [readthedocs](https://ex
 
 The software has been extensively validated against the Ariel radiometric model ArielRad [@arielrad], the time domain simulator ExoSim [@exosim] and custom simulations performed by the Ariel consortium. 
 ExoRad 2.0 is now used not only by the Ariel consortium but also by other missions, such as the balloon-borne NASA EXCITE mission [@excite], the space telescope Twinkle [@twinkle], and an adaptation for the James Webb Space Telescope [@jwst] is under preparation.
-
+*Such JWST has been tested against the JWST Exposure Time Calculator [@pandeia] returning consistent results as an ulterior validation for the ExoRad code.*
+*Although the code has been validated and used mostly for space and airborne-based telescopes, there are no indications of any limitations to adaptation for ground-based telescopes as well.*
 # ExoRad 2 features
 
 ExoRad 2.0 is a simulator able to accurately predict the telescope performance in observing a candidate target for all the mission photometric and spectroscopic channels. The software inputs are a target description and a parameterization of the instrument. The software parses the description of the instrument, and estimates the total optical efficiency, by combining the optical elements and the foregrounds (defined as any optical layer between the target star and the telescope aperture). Then it combines the optical efficiency with the detector quantum efficiency to obtain the photon conversion efficiency (see \autoref{fig:efficiency}). For the target, the software estimates the flux at the telescope aperture by parsing the source description: at the moment of writing the software is compatible with black body sources, Phoenix stellar spectra [@phoenix] or custom files describing the source spectral energy density versus wavelength. Then ExoRad 2.0 propagates the source flux through the foregrounds and the telescope optical path, estimating the total flux at the focal plane. Similarly, the software estimates the diffuse light contributions from the zodiacal emission, from any user-defined foregrounds between the source and the telescope aperture (e.g. the Earth's atmosphere), and from the self-emission of each optical element of the instrument.
