### 2023-06-22 #5348 @dfm
Thanks @skendrew and @eas342 for your reviews and suggestions here!! I agree that adding a worked example as suggested to the documentation would increase the impact and usability of the code.

@LorenzoMugnai — Let us know if you have any questions with regards to these final suggestions, and let's get this over the finish line!
