### 2023-07-06 #5348 @LorenzoMugnai
Hi all, sorry for the late reply.
I just updated the manuscript, and I will ask the bot to compile it in a minute, but first I will try now to address all your comments.

`- I think it would be very useful if you could include in the text of the paper a short discussion of the use of the code for ground-based vs. space-based instruments. Or perhaps include an example in the code for a ground-based telescope/instrument? I gather it's been written primarily for space-based (or airborne) systems, are there limitations to its use for ground-based systems?`

We are not aware of any limitations for a ground-based system. We didn't have the chance to apply the simulator to such a use case yet, but given the possibility to add multiple foregrounds, such as the atmospheric emission and transmission, we don't foresee any problem in implementing a ground-based telescope on ExoRad. We now mention it in the manuscript.

`- I can see in the payload_example input file that there are several options for providing a PSF. The description in the paper and documentation of this is fairly sparse. Can you provide a bit more information on what options are available for providing custom PSFs? Are time-varying PSFs supported?`

You are right: we didn't document properly the PSF module. We now included a dedicated section in the docs. However, time-varying PSFs are not supported by this code, as ExoRad is a Radiometric simulator and it cannot support time-dependent objects.

`- has the code been validated against any actual working telescopes/instruments? you mention validation against ArielRad and ExoSim; can you elaborate in the paper on how these have been validated against working systems? this would be useful for the reader.`

Sure, ExoSim has been validated against HST and JWST, and therefore our validation against ExoSim code is validation against these systems.
However, in the paper, we now mention the validation against Pandexo (which is validation against JWST). We prefer this example because the Pandexo software is widely known and it is more similar to ExoRad than ExoSim and therefore is a like-to-like comparison. We didn't include a full example with all the configuration file in the documentation because a dedicated version of ExoRad for the JWST (which will support different observing modes) is under preparation, and we plan to present it with a dedicated publication.

`- it would be good to clarify what the min/max wavelengths are that the code can handle. I can see that the user can specify cutoff wavelengths, but presumably, there is some limit. `

From a software perspective, we are not aware of any limitation on wavelength range: ExoRad functions should be able to handle any wavelength range.
Cutoff wavelengths are related to detectors: different detectors have different responsivities, which can be described by the quantum efficiencies data and by the detector cutoff.
So, the only limits we can think about are related to the physics, and not to the software. For example, if we want to simulate a microwave or an X-ray observation, we might need to update the code with a different instrument than a photometer and spectrometer, because the detectors used are different. This however is not, in our opinion, a software limitation, but a use case.

Please, let me know if you have any more questions or if my answers are not enough.
Thanks again for your help.
