### 2023-08-11 0978bb0
## [2.1.120] - 2023-08-10 ### Fixed - added support for np.int32 for Windows by @wqcannon - broken pyYaml package 5.4 with cython 3. Moved back to 5.3.1
