### 2023-09-21 9550d45
Small edits for JOSS review
