+++ joss/paper.bib
@@ -51,7 +51,7 @@ archivePrefix = {arXiv},
 @ARTICLE{pandexo,
        author = {{Batalha}, Natasha E. and {Mandell}, Avi and {Pontoppidan}, Klaus and {Stevenson}, Kevin B. and {Lewis}, Nikole K. and {Kalirai}, Jason and {Earl}, Nick and {Greene}, Thomas and {Albert}, Lo{\"\i}c and {Nielsen}, Louise D.},
         title = "{PandExo: A Community Tool for Transiting Exoplanet Science with JWST \& HST}",
-      journal = {\pasp},
+      journal = {Publications of the ASP},
      keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
          year = 2017,
         month = jun,
@@ -69,7 +69,7 @@ archivePrefix = {arXiv},
 @ARTICLE{phoenix,
        author = {{Baraffe}, Isabelle and {Homeier}, Derek and {Allard}, France and {Chabrier}, Gilles},
         title = "{New evolutionary models for pre-main sequence and main sequence low-mass stars down to the hydrogen-burning limit}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {stars: evolution, stars: low-mass, stars: pre-main sequence, Hertzsprung-Russell and C-M diagrams, convection, Astrophysics - Solar and Stellar Astrophysics},
          year = 2015,
         month = may,
@@ -122,7 +122,7 @@ archivePrefix = {arXiv},
 @ARTICLE{jwst,
        author = {{Gardner}, Jonathan P. and {Mather}, John C. and {Clampin}, Mark and {Doyon}, Rene and {Greenhouse}, Matthew A. and {Hammel}, Heidi B. and {Hutchings}, John B. and {Jakobsen}, Peter and {Lilly}, Simon J. and {Long}, Knox S. and {Lunine}, Jonathan I. and {McCaughrean}, Mark J. and {Mountain}, Matt and {Nella}, John and {Rieke}, George H. and {Rieke}, Marcia J. and {Rix}, Hans-Walter and {Smith}, Eric P. and {Sonneborn}, George and {Stiavelli}, Massimo and {Stockman}, H.~S. and {Windhorst}, Rogier A. and {Wright}, Gillian S.},
         title = "{The James Webb Space Telescope}",
-      journal = {\ssr},
+      journal = {Space Science Reviews},
      keywords = {galaxies: formation, infrared: general, planetary systems, space vehicles: instruments, stars: formation, Astrophysics},
          year = 2006,
         month = apr,
@@ -142,7 +142,7 @@ archivePrefix = {arXiv},
 @ARTICLE{rauscher,
[…]
