### 2023-09-21 #5348 @dfm
@LorenzoMugnai — I've opened a small PR with some minor edits to the manuscript, please take a look and merge or let me know what you think.

Once you've done that:
