+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = exorad
-version = 2.1.120
+version = 2.1.121
 author = Lorenzo V. Mugnai, Enzo Pascale
 author_email = lorenzo.mugnai@uniroma1.it
 description = The generic point source radiometric model
