### 2023-09-21 #5348 @LorenzoMugnai
Hello everybody, thank you very much for your work! I really appreciated your feedbacks.

@dfm I merged your request and followed your instructions:

1. looks good to me
2. v2.1.121(https://github.com/ExObsSim/ExoRad2-public/releases/tag/v2.1.121)
3. 10.5281/zenodo.8367214 (https://zenodo.org/record/8367214)
