### 2023-09-26 #5348 @dfm
@skendrew, @eas342 — Many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

The paper isn't yet rendering properly on the publication page which sometimes happens because of a server-side caching issue. I think this normally resolves within the next few hours, but let's check back then so double check.
